rag_based_llm_auichat/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.8.1
annotated-types==0.7.0
anyio==4.8.0
arrow==1.3.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.0.1
beautifulsoup4==4.13.3
bigtree==0.25.2
binaryornot==0.4.4
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.3
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cookiecutter==2.6.0
cryptography==44.0.2
cycler==0.12.1
dask==2025.3.0
dataclasses-json==0.6.7
decorator==5.2.1
Deprecated==1.2.18
dill==0.3.9
dirtyjson==1.0.8
distro==1.9.0
docker==7.1.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.8
feast==0.47.0
ffmpy==0.5.0
filelock==3.18.0
filetype==1.2.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
gradio==5.22.0
gradio_client==1.8.0
greenlet==3.1.1
groovy==0.1.2
grpcio==1.71.0
grpcio-tools==1.71.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
ipinfo==5.1.1
ipython==8.34.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
llama-cloud==0.1.14
llama-index==0.12.24
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.24.post1
llama-index-embeddings-huggingface==0.5.2
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.9
llama-index-llms-huggingface-api==0.4.1
llama-index-llms-openai==0.3.25
llama-index-llms-huggingface==0.3.0
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.6
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-qdrant==0.4.3
llama-parse==0.5.8
locket==1.0.0
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.1.0
mypy==1.15.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.66.3
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
partd==1.4.2
passlib==1.7.4
pexpect==4.9.0
pillow==11.1.0
portalocker==2.10.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.8.1
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.7.0
PyMySQL==1.1.1
pyparsing==3.2.1
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.1
PyYAML==6.0.2
qdrant-client==1.13.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
rpds-py==0.23.1
ruff==0.11.2
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
secure==0.3.0
semantic-version==2.10.0
sentence-transformers==3.4.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.39
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.18
stack-data==0.6.3
starlette==0.45.3
striprtf==0.0.26
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.9.0
tldextract==5.1.3
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
toolz==1.0.0
torch==2.6.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
triton==3.2.0
typeguard==4.4.2
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
uvicorn-worker==0.3.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
widgetsnbextension==4.0.13
wrapt==1.17.2
yarl==1.18.3
zenml==0.75.0
zipp==3.21.0
google-cloud-storage>=2.8.0
google-cloud-aiplatform>=1.25.0
google-auth>=2.15.0
